Refactoring cct-service and interop-service into letters-service
